move check
